save last seen binance heights and signed height

+++ x/swapservice/module.go
@@ -105,7 +105,7 @@ func (am AppModule) NewQuerierHandler() sdk.Querier {
 
 func (am AppModule) BeginBlock(ctx sdk.Context, req abci.RequestBeginBlock) {
 	ctx.Logger().Debug("Begin Block", "height", req.Header.Height)
-	am.txOutStore.NewBlock(req.Header.Height)
+	am.txOutStore.NewBlock(uint64(req.Header.Height))
 }
 
 func (am AppModule) EndBlock(ctx sdk.Context, req abci.RequestEndBlock) []abci.ValidatorUpdate {
